+++ etils/ecolab/auto_display_utils.py
@@ -30,6 +30,7 @@ from etils.ecolab import highlight_util
 from etils.ecolab.inspects import core as inspects
 from etils.etree import jax as etree  # pylint: disable=g-importing-member
 import IPython
+from IPython.lib import pretty
 import packaging.version
 
 _T = TypeVar('_T')
@@ -434,14 +435,14 @@ def _display_and_return(
   if _Options.SPEC in options:  # Convert to spec
     x = etree.spec_like(x)
 
-  repr_fn = repr
+  repr_fn = pretty.pretty
   display_fn = IPython.display.display
   if line_code and _Options.SYNTAX_HIGHLIGHT not in options:
     print(line_code + ' = ', end='')
     # When the next element is a `IPython.display`, the next element is
     # displayed on a new line. This is because `display()` create a new
     # <div> section. So use standard `print` when line is displayed.
-    display_fn = lambda x: print(repr(x))
+    display_fn = lambda x: print(pretty.pretty(x))
 
   if _Options.PPRINT in options:
     repr_fn = epy.pretty_repr